Fix #61, 62: very frequent freezing of VM during teardown
This will only mitigate the issue and doesn't fix the root cause. The
action doesn't shutdown the VM anymore. Since the action is run
inside a VM itself, everything will be cleaned up automatically.
Hopefully this will make the issue less likely to occur.

+++ src/action/action.ts
@@ -119,12 +119,16 @@ export class Action {
       } finally {
         core.startGroup('Tearing down VM')
         await this.syncBack(vm.ipAddress)
-        await vm.stop()
+
+        if (this.input.shutdownVm) {
+          await vm.stop()
+        }
       }
     } finally {
       try {
-        await vm.terminate()
-        fs.rmSync(this.tempPath, {recursive: true})
+        if (this.input.shutdownVm) {
+          await vm.terminate()
+        }
       } finally {
         core.endGroup()
       }
@@ -168,6 +172,7 @@ export class Action {
       hypervisorDirectory,
       resourcesDirectory,
       firmwareDirectory,
+      this.input,
       {
         ...config,
         diskImage: path.join(resourcesDirectory, this.targetDiskName),
